+++ dao/dao.GetObjectRevisionsByUser.go
@@ -9,13 +9,13 @@ import (
 
 // GetObjectRevisionsByUser retrieves a list of revisions for an object.
 func (dao *DataAccessLayer) GetObjectRevisionsByUser(
-	user models.ODUser, pagingRequest PagingRequest, object models.ODObject, checkACM CheckACM) (models.ODObjectResultset, error) {
+	user models.ODUser, pagingRequest PagingRequest, object models.ODObject, withProperties bool) (models.ODObjectResultset, error) {
 	tx, err := dao.MetadataDB.Beginx()
 	if err != nil {
 		dao.GetLogger().Error("Could not begin transaction", zap.String("err", err.Error()))
 		return models.ODObjectResultset{}, err
 	}
-	response, err := getObjectRevisionsByUserInTransaction(tx, user, pagingRequest, object, checkACM)
+	response, err := getObjectRevisionsByUserInTransaction(tx, user, pagingRequest, object, withProperties)
 	if err != nil {
 		dao.GetLogger().Error("Error in GetObjectRevisionsByUser", zap.String("err", err.Error()))
 		tx.Rollback()
@@ -25,7 +25,7 @@ func (dao *DataAccessLayer) GetObjectRevisionsByUser(
 	return response, err
 }
 
-func getObjectRevisionsByUserInTransaction(tx *sqlx.Tx, user models.ODUser, pagingRequest PagingRequest, object models.ODObject, checkACM CheckACM) (models.ODObjectResultset, error) {
+func getObjectRevisionsByUserInTransaction(tx *sqlx.Tx, user models.ODUser, pagingRequest PagingRequest, object models.ODObject, withProperties bool) (models.ODObjectResultset, error) {
 	response := models.ODObjectResultset{}
 	// NOTE: distinct is unfortunately used here because object_permission
 	// allows multiple records per object and grantee.
@@ -88,18 +88,20 @@ func getObjectRevisionsByUserInTransaction(tx *sqlx.Tx, user models.ODUser, pagi
 	response.PageRows = len(response.Objects)
 	response.PageCount = GetPageCount(response.TotalRows, response.PageSize)
 
-	// Redact by ACM access and set permissions
+	// Get detail information for each object
 	permissions := []models.ODObjectPermission{}
-	for i, o := range response.Objects {
-		ok := checkACM(&o)
-		if !ok {
-			//Preserve the id field and list the changeCount field
-			id := response.Objects[i].ID
-			response.Objects[i] = models.ODObject{}
-			response.Objects[i].ID = id
-			response.Objects[i].ChangeCount = -1
+	for i := 0; i < len(response.Objects); i++ {
+		// Populate properties if requested
+		if withProperties {
+			properties, err := getPropertiesForObjectRevisionInTransaction(tx, response.Objects[i])
+			if err != nil {
+				return response, err
+			}
+			response.Objects[i].Properties = properties
 		}
+		// Permissions
 		if len(permissions) == 0 {
+			// Not yet retrieved, do it now
 			permissions, err = getPermissionsForObjectInTransaction(tx, object)
 			if err != nil {
 				return response, err
@@ -107,6 +109,5 @@ func getObjectRevisionsByUserInTransaction(tx *sqlx.Tx, user models.ODUser, pagi
 		}
 		response.Objects[i].Permissions = permissions
 	}
-
 	return response, err
 }