+++ dao/fake.go
@@ -156,12 +156,7 @@ func (fake *FakeDAO) GetObjectRevision(object models.ODObject, loadProperties bo
 }
 
 // GetObjectRevisionsByUser for FakeDAO
-func (fake *FakeDAO) GetObjectRevisionsByUser(user models.ODUser, pagingRequest PagingRequest, object models.ODObject, checkACM CheckACM) (models.ODObjectResultset, error) {
-	return fake.ObjectResultSet, fake.Err
-}
-
-// GetObjectRevisionsWithPropertiesByUser for FakeDAO
-func (fake *FakeDAO) GetObjectRevisionsWithPropertiesByUser(user models.ODUser, pagingRequest PagingRequest, object models.ODObject, checkACM CheckACM) (models.ODObjectResultset, error) {
+func (fake *FakeDAO) GetObjectRevisionsByUser(user models.ODUser, pagingRequest PagingRequest, object models.ODObject, withProperties bool) (models.ODObjectResultset, error) {
 	return fake.ObjectResultSet, fake.Err
 }