+++ changelog.md
@@ -14,6 +14,7 @@ FORMAT: 1A
 * FIX: Init script now checks that paths are configured as absolute.
 * ENH: Init script now uses logging with log levels.
 * ENH: Server startup will now block forever until main ZK cluster is reachable.
+* REF: Internal refactor for retrieving object revisions.
 
 ## Release v1.0.1.17
 --------------------